app/requirements.txt
Flask==2.3.2
Flask-Session2==1.3.1
passlib==1.7.4
peewee==3.16.2
peewee-migrate==1.7.1
psycopg2-binary==2.9.6
pytz==2022.7.1
redis==4.5.4
requests==2.32.4
uwsgidecorators==1.1.0
Flask-SocketIO==5.6.0
uvicorn==0.40.0
python-socketio==5.16.0